Rewrite resource paths, base url, typos
Added is a series of rules which will properly route
all requests to the apache server for resources located
in the static/ (resources) directory to the proper location.

In addition, the base url (/) will direct to the ontodocs index
page, and multiple mentions of 'server' are now properly worded as
'serve'.

Signed-off-by: TJ Balon <tbalon@mitre.org>